stresstest tests and reset fork uncomment
