+++ roles_royce/applications/panic_button_app/config/dao_config_builder.py
@@ -6,7 +6,7 @@ from defabipedia.types import Chain, Blockchain
 
 
 def main(dao: DAO, blockchain: Blockchain, balancer: list[BalancerPosition], aura: list[AuraPosition], lido: list[LidoPosition]):
-    PUBLIC_ETH_NODE_URL = 'https://rpc.mevblocker.io'
+    PUBLIC_ETH_NODE_URL = 'https://eth.llamarpc.com'
     PUBLIC_GC_NODE_URL = 'https://rpc.ankr.com/gnosis'
 
     w3_eth = Web3(Web3.HTTPProvider(os.environ.get("RR_ETH_FORK_URL", PUBLIC_ETH_NODE_URL)))
