+++ tests/applications/panic_button_app/test_stresstest_gno_gno.json
@@ -1,0 +1,71 @@
+{
+    "dao": "GnosisDAO",
+    "blockchain": "gnosis",
+    "general_parameters": [
+        {
+            "name": "percentage",
+            "label": "Percentage",
+            "type": "input",
+            "rules": {
+                "min": 0.001,
+                "max": 100
+            }
+        }
+    ],
+    "positions": [
+        {
+            "protocol": "Balancer",
+            "position_id": "107",
+            "position_id_tech": "0xa99FD9950B5D5dCeEaf4939E221dcA8cA9B938aB",
+            "position_id_human_readable": "gnosis_Balancer_WETH_BAL_GNO_WXDAI",
+            "exec_config": [
+                {
+                    "function_name": "exit_1_1",
+                    "label": "Withdraw (proportional)",
+                    "test": true,
+                    "stresstest": false,
+                    "description": "Exit pool on Balancer with proportional exit. (Not for recovery mode)",
+                    "parameters": [
+                        {
+                            "name": "bpt_address",
+                            "type": "constant",
+                            "value": "0xa99FD9950B5D5dCeEaf4939E221dcA8cA9B938aB"
+                        },
+                        {
+                            "name": "max_slippage",
+                            "label": "Max slippage",
+                            "type": "input",
+                            "rules": {
+                                "min": 0.001,
+                                "max": 100
+                            }
+                        }
+                    ]
+                },
+                {
+                    "function_name": "exit_1_3",
+                    "label": "Withdraw (proportional) (Recovery mode)",
+                    "test": true,
+                    "stresstest": false,
+                    "description": "Exit pool on Balancer with proportional exit. (Recovery Mode)",
+                    "parameters": [
+                        {
+                            "name": "bpt_address",
+                            "type": "constant",
+                            "value": "0xa99FD9950B5D5dCeEaf4939E221dcA8cA9B938aB"
+                        },
+                        {
+                            "name": "max_slippage",
+                            "label": "Max slippage",
+                            "type": "input",
+                            "rules": {
+                                "min": 0.001,
+                                "max": 100
+                            }
+                        }
+                    ]
+                }
+            ]
+        }
+    ]
+}
