--- tests/applications/panic_button_app/test_stresstest_v2.json
@@ -1,68 +1,0 @@
-{
-    "position_id": "33",
-    "position_id_tech": "0xDd1fE5AD401D4777cE89959b7fa587e569Bf125D",
-    "position_id_human_readable": "ethereum_Aura_rETH_WETH",
-    "protocol": "Aura",
-    "exec_config": [
-        {
-            "function_name": "exit_1",
-            "label": "Unstake",
-            "test": true,
-            "stresstest": false,
-            "description": "Unstake BPT from Aura gauge",
-            "parameters": [
-                {
-                    "name": "rewards_address",
-                    "type": "constant",
-                    "value": "0xDd1fE5AD401D4777cE89959b7fa587e569Bf125D"
-                }
-            ]
-        },
-        {
-            "function_name": "exit_2_1",
-            "label": "Unstake + withdraw (proportional)",
-            "test": true,
-            "stresstest": false,
-            "description": "Unstake the BPT from Aura and exit pool on Balancer with proportional exit (Not for recovery mode)",
-            "parameters": [
-                {
-                    "name": "rewards_address",
-                    "type": "constant",
-                    "value": "0xDd1fE5AD401D4777cE89959b7fa587e569Bf125D"
-                },
-                {
-                    "name": "max_slippage",
-                    "label": "Max Slippage",
-                    "type": "input",
-                    "rules": {
-                        "min": 0.001,
-                        "max": 100
-                    }
-                }
-            ]
-        },
-        {
-            "function_name": "exit_2_3",
-            "label": "Unstake + withdraw (proportional) (Recovery mode)",
-            "test": true,
-            "stresstest": false,
-            "description": "Unstake the BPT from Aura and exit pool on Balancer with proportional exit. (Recovery mode)",
-            "parameters": [
-                {
-                    "name": "rewards_address",
-                    "type": "constant",
-                    "value": "0xDd1fE5AD401D4777cE89959b7fa587e569Bf125D"
-                },
-                {
-                    "name": "max_slippage",
-                    "label": "Max Slippage",
-                    "type": "input",
-                    "rules": {
-                        "min": 0.001,
-                        "max": 100
-                    }
-                }
-            ]
-        }
-    ]
-}
